Move dsync out from sso
