+++ src/dsync/DirectoryList/index.lite.tsx
@@ -1,12 +1,12 @@
 import { useStore, Show, onUpdate } from '@builder.io/mitosis';
 import { DirectorySyncProviders, type Directory } from '../types';
-import LoadingContainer from '../../../shared/LoadingContainer/index.lite';
+import LoadingContainer from '../../shared/LoadingContainer/index.lite';
 import type { DirectoryListProps } from '../types';
 import defaultClasses from './index.module.css';
-import cssClassAssembler from '../../utils/cssClassAssembler';
-import Table from '../../../shared/Table/index.lite';
-import { BadgeProps, TableProps } from '../../../shared/types';
-import EmptyState from '../../../shared/EmptyState/index.lite';
+import cssClassAssembler from '../../sso/utils/cssClassAssembler';
+import Table from '../../shared/Table/index.lite';
+import { BadgeProps, TableProps } from '../../shared/types';
+import EmptyState from '../../shared/EmptyState/index.lite';
 
 const DEFAULT_VALUES = {
   directoryListData: [] as Directory[],
