+++ src/dsync/index.ts
@@ -1,0 +1,4 @@
+export { default as CreateDirectory } from './CreateDirectory/index.lite';
+export { default as DirectoryList } from './DirectoryList/index.lite';
+export { default as EditDirectory } from './EditDirectory/index.lite';
+export { default as DirectoriesWrapper } from './DirectoriesWrapper/index.lite';
