+++ src/sso/index.ts
@@ -6,7 +6,3 @@ export { default as ConnectionList } from './connections/ConnectionList/index.li
 export { default as EditOIDCConnection } from './connections/EditConnection/oidc/index.lite';
 export { default as EditSAMLConnection } from './connections/EditConnection/saml/index.lite';
 export { default as ConnectionsWrapper } from './connections/ConnectionsWrapper/index.lite';
-export { default as CreateDirectory } from './dsync/CreateDirectory/index.lite';
-export { default as DirectoryList } from './dsync/DirectoryList/index.lite';
-export { default as EditDirectory } from './dsync/EditDirectory/index.lite';
-export { default as DirectoriesWrapper } from './dsync/DirectoriesWrapper/index.lite';
